#778 Moved Cursor3D into a file of its own.
